% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bgzip}
\alias{bgzip}
\title{Bgzips a VCF file}
\usage{
bgzip(
  bin_path = system.file("htslib/bin", "bgzip", package = "RandomCode"),
  file = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{Required} Path to bcftools binary. Default tools/htslib/tabix.}

\item{file}{\link{Required} Path to VCF file to bgzip.}

\item{verbose}{\link{Optional} Enables progress messages. Default False.}
}
\description{
This function takes a VCF file and Bgzips it
}
