% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_variants.R
\name{call_sv_manta}
\alias{call_sv_manta}
\title{Call Somatic/Germline structural variants using Manta
This function takes a pair of matched samples or single | multiple germline samples
and calls variants on them. Variant calling mode is established based on wheather tumor_bam is supplied or not.}
\usage{
call_sv_manta(
  bin_path = system.file("tools/manta/bin", "configManta.py", package = "RandomCode"),
  tumor_bam = "",
  normal_bam = "",
  ref_genome = "",
  output_dir = "",
  patient_id = "",
  verbose = FALSE,
  targeted = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to strelka binary. Somatic or Germline.}

\item{tumor_bam}{\link{REQUIRED} Path to tumor  bam file.}

\item{normal_bam}{\link{OPTIONAL} Path to normal samples bam files.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{patient_id}{\link{OPTIONAL} Patient ID to name files after. If not given tumor_bam file name will be used, and if this is not given then normal_bam file will be used.}

\item{verbose}{\link{DEFAULT==FALSE} Enables progress messages.}

\item{targeted}{\link{OPTIONAL} If exome/capture method. Default FALSE}

\item{threads}{\link{OPTIONAL} Number of threads per job. Default 3}
}
\description{
Call Somatic/Germline structural variants using Manta
This function takes a pair of matched samples or single | multiple germline samples
and calls variants on them. Variant calling mode is established based on wheather tumor_bam is supplied or not.
}
